+++ src/aellux.orchestrator.js
@@ -25,7 +25,7 @@ root.Aellux = Object.assign(AelluxForceUpdate, root.Aellux, {
         .forEach(extensionName => Aellux.ext(extensionName));
     }
 
-    await new Promise((resolve, reject) => {
+    await new Promise((resolve) => {
       const startUpdateCallback = function () {
         Aellux.update().then(function () {
           document.removeEventListener(